add test for testcase
