+++ testdata/event.json
@@ -1,0 +1,22 @@
+{
+    "version": "1.0",
+    "context": {
+        "eventType": "viewer-response"
+    },
+    "viewer": {
+        "ip": "1.2.3.4"
+    },
+    "request": {
+        "method": "GET",
+        "uri": "/index.html",
+        "headers": {},
+        "cookies": {},
+        "querystring": {}
+    },
+    "response": {
+        "statusCode": 200,
+        "statusDescription": "OK",
+        "headers": {},
+        "cookies": {}
+    }
+}
