+++ testdata/expect.jsonnet
@@ -1,0 +1,12 @@
+{
+  response: {
+    headers: {
+      'cache-control': {
+        value: 'public, max-age=6307200',
+      },
+    },
+    statusDescription: 'OK',
+    cookies: {},
+    statusCode: 200,
+  },
+}
