+++ testdata/expect.yaml
@@ -1,0 +1,8 @@
+---
+response:
+  headers:
+    cache-control:
+      value: public, max-age=6307200
+  statusDescription: OK
+  cookies: {}
+  statusCode: 200
